rename varrep and rep
