+++ src/callable/primitive/sum.rs
@@ -4,7 +4,7 @@ use crate::callable::core::*;
 use crate::error::*;
 use crate::internal_err;
 use crate::lang::*;
-use crate::object::rep::Rep;
+use crate::object::reptype::RepType;
 use crate::object::*;
 
 #[derive(Debug, Clone, PartialEq)]
@@ -20,7 +20,7 @@ impl Callable for PrimitiveSum {
         let (_, ellipsis) = self.match_arg_exprs(args, stack)?;
 
         if ellipsis.is_empty() {
-            return EvalResult::Ok(Obj::Vector(Vector::from(Rep::from(vec![0.0]))));
+            return EvalResult::Ok(Obj::Vector(Vector::from(RepType::from(vec![0.0]))));
         }
 
         let objects: Vec<Obj> = force_promises(ellipsis, stack)?
@@ -58,8 +58,8 @@ impl Callable for PrimitiveSum {
                                 for x in repr.inner().borrow().iter() {
                                     match *x {
                                         OptionNA::NA => {
-                                            let rep: Rep<OptionNA<f64>> =
-                                                Rep::from(vec![OptionNA::NA]);
+                                            let rep: RepType<OptionNA<f64>> =
+                                                RepType::from(vec![OptionNA::NA]);
                                             return EvalResult::Ok(Obj::Vector(Vector::from(rep)));
                                         }
                                         OptionNA::Some(x) => sum += x as i32 as f64,
@@ -70,8 +70,8 @@ impl Callable for PrimitiveSum {
                                 for x in repr.inner().borrow().iter() {
                                     match *x {
                                         OptionNA::NA => {
-                                            let rep: Rep<OptionNA<f64>> =
-                                                Rep::from(vec![OptionNA::NA]);
+                                            let rep: RepType<OptionNA<f64>> =
+                                                RepType::from(vec![OptionNA::NA]);
                                             return EvalResult::Ok(Obj::Vector(Vector::from(rep)));
                                         }
                                         OptionNA::Some(x) => sum += x as f64,
@@ -82,8 +82,8 @@ impl Callable for PrimitiveSum {
                                 for x in repr.inner().borrow().iter() {
                                     match *x {
                                         OptionNA::NA => {
-                                            let rep: Rep<OptionNA<f64>> =
-                                                Rep::from(vec![OptionNA::NA]);
+                                            let rep: RepType<OptionNA<f64>> =
+                                                RepType::from(vec![OptionNA::NA]);
                                             return EvalResult::Ok(Obj::Vector(Vector::from(rep)));
                                         }
                                         OptionNA::Some(x) => sum += x,
@@ -96,7 +96,7 @@ impl Callable for PrimitiveSum {
                     _ => return internal_err!(),
                 }
             }
-            EvalResult::Ok(Obj::Vector(Vector::from(Rep::from(vec![sum]))))
+            EvalResult::Ok(Obj::Vector(Vector::from(RepType::from(vec![sum]))))
         } else {
             let mut sum: i32 = 0;
 
@@ -108,8 +108,8 @@ impl Callable for PrimitiveSum {
                                 for x in repr.inner().borrow().iter() {
                                     match *x {
                                         OptionNA::NA => {
-                                            let rep: Rep<OptionNA<i32>> =
-                                                Rep::from(vec![OptionNA::NA]);
+                                            let rep: RepType<OptionNA<i32>> =
+                                                RepType::from(vec![OptionNA::NA]);
                                             return EvalResult::Ok(Obj::Vector(Vector::from(rep)));
                                         }
                                         OptionNA::Some(x) => sum += x as i32,
@@ -120,8 +120,8 @@ impl Callable for PrimitiveSum {
                                 for x in repr.inner().borrow().iter() {
                                     match *x {
                                         OptionNA::NA => {
-                                            let rep: Rep<OptionNA<i32>> =
-                                                Rep::from(vec![OptionNA::NA]);
+                                            let rep: RepType<OptionNA<i32>> =
+                                                RepType::from(vec![OptionNA::NA]);
                                             return EvalResult::Ok(Obj::Vector(Vector::from(rep)));
                                         }
                                         OptionNA::Some(x) => sum += x,
@@ -134,7 +134,7 @@ impl Callable for PrimitiveSum {
                     _ => return internal_err!(),
                 }
             }
-            EvalResult::Ok(Obj::Vector(Vector::from(Rep::from(vec![sum]))))
+            EvalResult::Ok(Obj::Vector(Vector::from(RepType::from(vec![sum]))))
         }
     }
 }
