+++ src/object/vector/mod.rs
@@ -7,8 +7,8 @@
 pub mod coercion;
 pub mod iterators;
 pub mod rep;
+pub mod reptype;
 pub mod types;
-pub mod vrep;
 
 mod subsets;
 pub use subsets::*;
